Ensure that it'll work
We're out in the weeds, friends. Sometimes, you have to verify that
something will work, because the fancy type features might not actually
work out when you go to implement them.

The ambiguous types in this concerned me, so I made sure that I could
grab a list of strings from the constraints.